requirements.txt
paho-mqtt==1.6.1
Flask==2.3.2